Add basic support for command groups
Adds support for registering a command group instead of a command. A command group is basically a dictionary of dictionaries or callables if the value at a key contains a dictionary then it is a sub group. If the value at a key contains a callable then it is a subcommand